backend/requirements.txt
asn1crypto==0.24.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cryptography==2.3
Flask==1.0.2
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
pycparser==2.18
PyMySQL==0.9.2
PyYAML==3.13
requests==2.19.1
six==1.11.0
SQLAlchemy==1.2.10
urllib3==1.23
Werkzeug==0.14.1
